Extract extends from parent configuration and migrate to model action
